+++ lib/Menu/viewInvitation.dart
@@ -95,9 +95,30 @@ class _viewInvitationState extends State<viewInvitation> {
                                     child: Text("Delete"),
                                     color: Colors.orange,
                                     onPressed: () async {
-                                      var res = await db.deleteInvitation(
-                                          invitationList[index].I_id);
-                                    }),
+                                  if (invitationList[index].user_id == '') {
+                                    showDialog(
+                                      context: context,
+                                      builder: (context) => AlertDialog(
+                                          title: Text("user Id is absense")),
+                                    );
+                                  } else {
+                                    var res = await db.deleteInvitation(
+                                       invitationList[index].I_id);
+                                    print("${res}ressss");
+
+                                    if (res == 200) {
+                                      showDialog(
+                                        context: context,
+                                        builder: (context) => AlertDialog(
+                                            title: Text("Successfully deleted ")),
+                                      );
+
+                                      print("success");
+                                    } else {
+                                      print("failure");
+                                    }
+                                  }
+                                },),
                               ),
                               Divider(
                                 color: Colors.greenAccent,
