+++ lib/Menu/viewItemOrder.dart
@@ -97,10 +97,31 @@ class _viewItemOrdersState extends State<viewItemOrders> {
                                 child: RaisedButton(
                                     child: Text("Delete"),
                                     color: Colors.orange,
-                                    onPressed: () async {
-                                      var res = await db.deleteMedicalItemsOrder(
-                                          itemOrderList[index].itmOrder_id);
-                                    }),
+                                     onPressed: () async {
+                                  if (itemOrderList[index].user_id == '') {
+                                    showDialog(
+                                      context: context,
+                                      builder: (context) => AlertDialog(
+                                          title: Text("user Id is absense")),
+                                    );
+                                  } else {
+                                    var res = await db.deleteMedicalItemsOrder(
+                                       itemOrderList[index].itmOrder_id);
+                                    print("${res}ressss");
+
+                                    if (res == 200) {
+                                      showDialog(
+                                        context: context,
+                                        builder: (context) => AlertDialog(
+                                            title: Text("Successfully deleted ")),
+                                      );
+
+                                      print("success");
+                                    } else {
+                                      print("failure");
+                                    }
+                                  }
+                                }),
                               ),
                               Divider(
                                 color: Colors.greenAccent,
